refactor audio_mixer
