+++ backend/engine/src/audio_mixer/mod.rs
@@ -1,3 +1,4 @@
-pub mod effects;
+pub mod control;
+pub mod volume;
 
-pub use effects::*;
+pub use control::*;
